# Faster state serialization