@Zireael07 2016-02-22
> Another problem we noticed is that the more edits get performed on a certain buffer the slower serializing its history becomes.

This might be the "slowdown after working for a while" problem some people are observing.