@mehcode 2016-02-22
I feel that preserving buffer history across restarts is a misfeature. I've never wanted to `undo` past when I've opened the editor and if I do accidentally it always comes with a "now where was I again.. this is weird" reaction.

With that said.. would it be possible to move history serialization into its own non-bundled package? That would keep core clean and allow whomever to pick up the work of making it "fast" later.