@alexandernst 2016-02-22
Preserving undo/redo history is a must-have for any editor. Keep in mind
that when you voluntarily close the editor, you're supposed to have
finished editing whatever you were editing, but accidental quits/crashes
are the other side of the story. And since no program ever made (and Atom
just won't be the exception) is crash-free, please keep the undo/redo
restore working and in-core.

2016-02-22 18:39 GMT+01:00 Ryan Leckey <notifications@github.com>:

> I feel that preserving buffer history across restarts is a misfeature.
> I've never wanted to undo past when I've opened the editor and if I do
> accidentally it always comes with a "now where was I again.. this is weird"
> reaction.
>
> With that said.. would it be possible to move history serialization into
> its own non-bundled package? That would keep core clean and allow whomever
> to pick up the work of making it "fast" later.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/atom/atom/pull/10925#issuecomment-187283468>.
>



-- 
alexandernst