@mehcode 2016-02-22
How about removing it from the normal serialization code path? Only serialize buffer history when we do #10605 (periodically save unsaved buffers) and only on an idle callback. I feel (as @alexandernst said) that this is more for a crash restore scenario and doing it on text editor change, all the time, is not ideal.