@steelbrain 2016-02-23
@as-cii Instead of enabling/disabling history entirely, wouldn't it be nice to have some sort of limit on it?